+++ apicast/src/backend/cache_handler.lua
@@ -1,7 +1,7 @@
 local setmetatable = setmetatable
 
 local _M = {
-  handlers = { }
+  handlers = setmetatable({}, { __index = { default = 'strict' } })
 }
 
 local mt = {
@@ -18,19 +18,37 @@ local mt = {
 }
 
 function _M.new(handler)
-  return setmetatable({ handler = handler }, mt)
+  local name = handler or _M.handlers.default
+  ngx.log(ngx.DEBUG, 'backend cache handler: ', name)
+  return setmetatable({ handler = name }, mt)
 end
-
 
 function _M.handlers.strict(cache, cached_key, response, ttl)
   if response.status == 200 then
-    ngx.log(ngx.INFO, 'apicast cache write key: ', cached_key, ', ttl: ', ttl )
-    cache:set(cached_key, 200, ttl or 0)
+    -- cached_key is set in post_action and it is in in authorize
+    -- so to not write the cache twice lets write it just in authorize
+    if ngx.var.cached_key ~= cached_key then
+      ngx.log(ngx.INFO, 'apicast cache write key: ', cached_key, ', ttl: ', ttl )
+      cache:set(cached_key, 200, ttl or 0)
+    end
+
     return true
   else
     ngx.log(ngx.NOTICE, 'apicast cache delete key: ', cached_key, ' cause status ', response.status)
     cache:delete(cached_key)
     return false, 'not authorized'
+  end
+end
+
+function _M.handlers.resilient(cache, cached_key, response, ttl)
+  local status = response.status
+
+  if status and status < 500 then
+    ngx.log(ngx.INFO, 'apicast cache write key: ', cached_key, ' status: ', status, ', ttl: ', ttl )
+
+    cache:set(cached_key, status, ttl or 0)
+
+    return status == 200
   end
 end